HUMAnN Tutorial with BiobakeryUtils.jl

🗒️ This tutorial is meant to be run in parallel with / mirror the official HUMAnN v3 tutorial

❓️ If you have questions about MetaPhlAn itself, please direct them to the bioBakery help forum

🤔 If you have questions about using the MetaPhlAn tools in julia, please open an issue, or start a discussion over on Microbiome.jl!

📔 For a function / type reference, jump to the bottom

Installation and setup

If you haven't already, check out the "Getting Started" page to install julia, create an environment and install BiobakeryUtils.jl, and hook up or install the MetaPhlAn v3 command line tools.

This tutorial assumes:

You are running julia v1.6 or greater

You have activated a julia Project that has BiobakeryUtils.jl installed

The humann python package is installed, and accessible from your PATH.

If any of those things aren't true, or you don't know if they're true, go back to "Getting Started" to see if you skipped a step. If you're still confused, please ask (see 3rd bullet point at the top)!

HUMAnN Databases

HUMAnN requires a number of specialized databases to work correctly. When you first install it, it comes with some demo databases that are much smaller, but can be used to complete this tutorial. However, for actually running real data, you'll want to take the time to download them - they're BIG! See here for more information.

For now, the easiest way to do this for now is via the shell, which you can access from the julia REPL by typing ;:

shell> humann_databases --help
usage: humann_databases [-h] [--available]
                        [--download <database> <build> <install_location>]
                        [--update-config {yes,no}]
                        [--database-location DATABASE_LOCATION]

HUMAnN Databases

optional arguments:
  -h, --help            show this help message and exit
  --available           print the available databases
  --download <database> <build> <install_location>
                        download the selected database to the install location
  --update-config {yes,no}
                        update the config file to set the new database as the default [DEFAULT: yes]
  --database-location DATABASE_LOCATION
                        location (local or remote) to pull the database

shell> humann_databases --available
HUMANnN2 Databases ( database : build = location )
chocophlan : full = http://huttenhower.sph.harvard.edu/humann_data/chocophlan/full_chocophlan.v296_201901b.tar.gz
chocophlan : DEMO = http://huttenhower.sph.harvard.edu/humann_data/chocophlan/DEMO_chocophlan.v296_201901b.tar.gz
uniref : uniref50_diamond = http://huttenhower.sph.harvard.edu/humann_data/uniprot/uniref_annotated/uniref50_annotated_v201901b_ful
l.tar.gz
uniref : uniref90_diamond = http://huttenhower.sph.harvard.edu/humann_data/uniprot/uniref_annotated/uniref90_annotated_v201901b_ful
l.tar.gz
uniref : uniref50_ec_filtered_diamond = http://huttenhower.sph.harvard.edu/humann_data/uniprot/uniref_ec_filtered/uniref50_ec_filte
red_201901b_subset.tar.gz
uniref : uniref90_ec_filtered_diamond = http://huttenhower.sph.harvard.edu/humann_data/uniprot/uniref_ec_filtered/uniref90_ec_filte
red_201901b_subset.tar.gz
uniref : DEMO_diamond = http://huttenhower.sph.harvard.edu/humann_data/uniprot/uniref_annotated/uniref90_DEMO_diamond_v201901b.tar.
gz
utility_mapping : full = http://huttenhower.sph.harvard.edu/humann_data/full_mapping_v201901b.tar.gz

For example, if you'd like to install these databases to /BigDrive/humann/, you could run

shell> humann_databases --download cholophlan full /BigDrive/humann/chocophlan
# ... lots of output

shell> humann_databases --download uniref uniref90_diamond /BigDrive/humann/uniref
# ... lots of output

shell> humann_databases --download utility_mapping full /BigDrive/humann/utility_mapping
# ... lots of output

At some point, I'll write some functions to automate this, but for now, doing this will update a configuration file, so you shouldn't have to worry about it again.

Running HUMAnN

Official tutorial link

Default outputs

Official tutorial link

load gene families file with humann_profile

look at contents

filter on stratified, unstratified

Manipulating tables

Official tutorial link

Do all of this in CommProfile

Normalize RPK to relative abundance

Official tutorial link

Regrouping genes to other functional categories

Official tutorial link

use humann_regroup

Attaching names to features

Official tutorial link

use humann_rename

HUMAnN for multiple samples

Official tutorial link

Plotting

Official tutorial link

Kevin should do this

Functions and types

BiobakeryUtils.humann — Function

humann(inputfile, output[, flags]; kwargs...)

Run humann command line tool on inputfile, putting outputs in output. Requires humann to be installed and accessible in the PATH (see Getting Started).

humann flag options (those that don't have a parameter) can be passed in an array, and other options can be passed with keyword arguments. For example, if on the command line you would run:

$ humann -i $INPUTFILE -o $OUTPUT --bypass-tranlated-search --input-format fastq.gz --output-format biom

using this function, you would write:

humann(INTPUTFILE, OUTPUT, ["bypass_translated_search"]; input_formal="fastq.gz", output_format="biom")

BiobakeryUtils.humann_profile — Method

Stratified import currently non-functional

BiobakeryUtils.humann_profiles — Method

Stratified import currently non-functional