MetaPhlAn Tutorial with BiobakeryUtils.jl

🗒️ This tutorial is meant to be run in parallel with / mirror the official MetaPhlAn v3 tutorial

❓️ If you have questions about MetaPhlAn itself, please direct them to the bioBakery help forum

🤔 If you have questions about using the MetaPhlAn tools in julia, please open an issue), or start a discussion over on Microbiome.jl)!

📔 For a function / type reference, jump to the bottom

Installation and setup

If you haven't already, check out the "Getting Started" page to install julia, create an environment and install BiobakeryUtils.jl, and hook up or install the MetaPhlAn v3 command line tools.

This tutorial assumes:

You are running julia v1.6 or greater

You have activated a julia Project that has BiobakeryUtils.jl installed

The metaphlan python package is installed, and accessible from your PATH.

If any of those things aren't true, or you don't know if they're true, go back to "Getting Started" to see if you skipped a step. If you're still confused, please ask (see 3rd bullet point at the top)!

Bowtie2 database

The first time you run metaphlan, it needs to download and unpack the marker database. If you don't care where this goes, don't worry about it - by default it will go into a subdirectory of your conda environment.

If your home folder has limited space, or you want to install it to a particular location (eg a faster drive), you can either pass the kewword argument bowtie2db="/path/to/location" to all metaphlan commands, or set the environment variable METAPHLAN_BOWTIE2_DB.

Input files

Official tutorial link

Some example files you can use to run this tutorial are available from the MetaPhlAn repo, and can be downloaded using the Downloads standard library in julia:

julia> using Downloads: download

julia> base_url = "https://github.com/biobakery/biobakery/raw/master/demos/biobakery_demos/data/metaphlan3/input/";

julia> files = [
    "SRS014476-Supragingival_plaque.fasta.gz",
    "SRS014494-Posterior_fornix.fasta.gz",
    "SRS014459-Stool.fasta.gz",
    "SRS014464-Anterior_nares.fasta.gz",
    "SRS014470-Tongue_dorsum.fasta.gz",
    "SRS014472-Buccal_mucosa.fasta.gz"
];

julia> for file in files
           download(joinpath(base_url, file), file)
       end

julia> readdir()
8-element Vector{String}:
 "Manifest.toml"
 "Project.toml"
 "SRS014459-Stool.fasta.gz"
 "SRS014464-Anterior_nares.fasta.gz"
 "SRS014470-Tongue_dorsum.fasta.gz"
 "SRS014472-Buccal_mucosa.fasta.gz"
 "SRS014476-Supragingival_plaque.fasta.gz"
 "SRS014494-Posterior_fornix.fasta.gz"

Run a single sample

Official tutorial link

For convenience, this package has the metaphlan() function, which can be used in your julia scripts to build and call the metaphlan command line tool.

For example, rather than call

$ metaphlan SRS014476-Supragingival_plaque.fasta.gz --input_type fasta > SRS014476-Supragingival_plaque_profile.txt

from the julia REPL, you can call

julia> metaphlan("SRS014476-Supragingival_plaque.fasta.gz", "SRS014476-Supragingival_plaque_profile.tsv"; input_type="fasta")ERROR: UndefVarError: metaphlan not defined

Output files

Official tutorial link

load output with metaphlan_profile

investigate with various functions (try to show similar things as tutorial)

Run on multiple cores

Official tutorial link

Run multiple samples

Official tutorial link

Merge outputs

Official tutorial link

use metaphlan_profile in loop and then commjoin

use metaphlan_profiles

Visualize results

Official tutorial link

Kevin should probably handle this part

Functions and Types

BiobakeryUtils.metaphlan — Method

metaphlan(inputfile, outputfile; kwargs...)

Run metaphlan command line tool on inputfile, creating output. Requires metaphlan to be installed and accessible in the PATH (see Getting Started).

metaphlan options can be passed via keyword arguments. For example, if on the command line you would run:

$ metaphlan some.fastq.gz output/some_profile.tsv --input_type fastq --nprocs 8

using this function, you would write:

metaphlan("some.fastq.gz", "output/some_profile.tsv"; input_type="fastq", nprocs=8)

Note: the input_type keyword is required.

Set the environmental variable "METAPHLANBOWTIE2DB" to specify the location where the markergene database is/will be installed, or pass bowtie2db = "some/path" as a keyword argument.

BiobakeryUtils.metaphlan_profile — Function

metaphlan_profile(path::AbstractString, level::Union{Int, Symbol}=:all; sample::AbstractString=basename(first(splitext(path))))

Compiles a MetaPhlAn file into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Sample name of the CommunityProfile can be specified by passing a sample argument. If name not given, the name of the file becomes the Sample name.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profile("test/files/metaphlan_single2.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 96 features in 1 samples

Feature names:
Bacteria, Archaea, Firmicutes...Ruminococcus_bromii, Bacteroides_vulgatus

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", 4)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 11 features in 1 samples

Feature names:
Clostridiales, Bacteroidales, Coriobacteriales...Firmicutes_unclassified, Pasteurellales

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 features in 1 samples

Feature names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus, sample = "sample2")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 features in 1 samples

Feature names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Sample names:
sample2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(paths::Array{<:AbstractString, 1}, level::Union{Int, Symbol}=:all)

Compiles MetaPhlAn profiles from multiple single tables into a CommunityProfile.

Examples ≡≡≡≡≡≡≡≡≡≡

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 129 features in 2 samples

Feature names:
Bacteria, Firmicutes, Bacteroidetes...Coprococcus_eutactus, Ruminococcus_bromii

Sample names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 46 features in 2 samples

Feature names:
Bacteroides, Roseburia, Faecalibacterium...Ruthenibacterium, Haemophilus

Sample names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], 5)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 24 features in 2 samples

Feature names:
Lachnospiraceae, Ruminococcaceae, Bacteroidaceae...Clostridiales_unclassified, Pasteurellaceae

Sample names:
metaphlan_single1, metaphlan_single2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(path::AbstractString, level::Union{Int, Symbol}=:all; keepunidentified=false)

Compiles MetaPhlAn profiles from a merged table into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Set keepunidentified flag to true to keep UNIDENTIFIED data.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 42 features in 7 samples

Feature names:
Archaea, Euryarchaeota, Methanobacteria...Actinomyces_viscosus, GCF_000175315

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 3 features in 7 samples

Feature names:
Methanobrevibacter, Methanosphaera, Actinomyces

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", 3)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 2 features in 7 samples

Feature names:
Methanobacteria, Actinobacteria

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 features in 7 samples

Feature names:
UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

# julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv", keepunidentified = true)
# CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 features in 7 samples

# Feature names:
# UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

# Sample names:
# sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

BiobakeryUtils.parsetaxa — Method

parsetaxa(taxstring::AbstractString; throw_error=true)

Given a string representing taxonmic levels as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria..."), separates taxonomic levels into elements of type Taxon in a vector.

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxa("k__Archaea|p__Euryarchaeota|c__Methanobacteria"; throw_error = true)
3-element Vector{Taxon}:
 Taxon("Archaea", :kingdom)
 Taxon("Euryarchaeota", :phylum)
 Taxon("Methanobacteria", :class)

BiobakeryUtils.parsetaxon — Method

parsetaxon(taxstring::AbstractString, taxlevel::Union{Int, Symbol})

Finds given taxonomic level in a string (as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria...")) and returns the clade and taxonomic level as a Taxon. If taxon level not given, function will return the most specific (lowest) taxonomic level available.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", 2)
Taxon("Euryarchaeota", :phylum)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", :kingdom)
Taxon("Archaea", :kingdom)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria")
Taxon("Methanobacteria", :class)