MetaPhlAn Tutorial with BiobakeryUtils.jl

🗒️ This tutorial is meant to be run in parallel with / mirror the official MetaPhlAn v3 tutorial

❓️ If you have questions about MetaPhlAn itself, please direct them to the bioBakery help forum

🤔 If you have questions about using the MetaPhlAn tools in julia, please open an issue, or start a discussion over on Microbiome.jl!

📔 For a function / type reference, jump to the bottom

Getting started

Installation of julia

Installation of metaphlan with Conda

Using an existing metaphlan installation

Input files

https://github.com/biobakery/biobakery/wiki/metaphlan3#input-files

Run a single sample

https://github.com/biobakery/biobakery/wiki/metaphlan3#run-a-single-sample

for now, use run(cmd)

Output files

https://github.com/biobakery/biobakery/wiki/metaphlan3#output-files

load output with metaphlan_profile

investigate with various functions (try to show similar things as tutorial)

Run on multiple cores

https://github.com/biobakery/biobakery/wiki/metaphlan3#run-on-multiple-cores

Run multiple samples

https://github.com/biobakery/biobakery/wiki/metaphlan3#run-multiple-samples

Merge outputs

https://github.com/biobakery/biobakery/wiki/metaphlan3#merge-outputs

use metaphlan_profile in loop and then commjoin

use metaphlan_profiles

Visualize results

https://github.com/biobakery/biobakery/wiki/metaphlan3#visualize-results

Kevin should probably handle this part

Functions and Types

BiobakeryUtils.metaphlan_profile — Function

metaphlan_profile(path::AbstractString, level::Union{Int, Symbol}=:all; sample::AbstractString=basename(first(splitext(path))))

Compiles a MetaPhlAn file into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Sample name of the CommunityProfile can be specified by passing a sample argument. If name not given, the name of the file becomes the Sample name.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profile("test/files/metaphlan_single2.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 96 features in 1 samples

Feature names:
Bacteria, Archaea, Firmicutes...Ruminococcus_bromii, Bacteroides_vulgatus

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", 4)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 11 features in 1 samples

Feature names:
Clostridiales, Bacteroidales, Coriobacteriales...Firmicutes_unclassified, Pasteurellales

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 features in 1 samples

Feature names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Sample names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus, sample = "sample2")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 features in 1 samples

Feature names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Sample names:
sample2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(paths::Array{<:AbstractString, 1}, level::Union{Int, Symbol}=:all)

Compiles MetaPhlAn profiles from multiple single tables into a CommunityProfile.

Examples ≡≡≡≡≡≡≡≡≡≡

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 129 features in 2 samples

Feature names:
Bacteria, Firmicutes, Bacteroidetes...Coprococcus_eutactus, Ruminococcus_bromii

Sample names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 46 features in 2 samples

Feature names:
Bacteroides, Roseburia, Faecalibacterium...Ruthenibacterium, Haemophilus

Sample names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], 5)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 24 features in 2 samples

Feature names:
Lachnospiraceae, Ruminococcaceae, Bacteroidaceae...Clostridiales_unclassified, Pasteurellaceae

Sample names:
metaphlan_single1, metaphlan_single2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(path::AbstractString, level::Union{Int, Symbol}=:all; keepunidentified=false)

Compiles MetaPhlAn profiles from a merged table into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Set keepunidentified flag to true to keep UNIDENTIFIED data.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 42 features in 7 samples

Feature names:
Archaea, Euryarchaeota, Methanobacteria...Actinomyces_viscosus, GCF_000175315

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 3 features in 7 samples

Feature names:
Methanobrevibacter, Methanosphaera, Actinomyces

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", 3)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 2 features in 7 samples

Feature names:
Methanobacteria, Actinobacteria

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 features in 7 samples

Feature names:
UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

Sample names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

# julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv", keepunidentified = true)
# CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 features in 7 samples

# Feature names:
# UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

# Sample names:
# sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

BiobakeryUtils.parsetaxa — Method

parsetaxa(taxstring::AbstractString; throw_error=true)

Given a string representing taxonmic levels as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria..."), separates taxonomic levels into elements of type Taxon in a vector.

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxa("k__Archaea|p__Euryarchaeota|c__Methanobacteria"; throw_error = true)
3-element Vector{Taxon}:
 Taxon("Archaea", :kingdom)
 Taxon("Euryarchaeota", :phylum)
 Taxon("Methanobacteria", :class)

BiobakeryUtils.parsetaxon — Method

parsetaxon(taxstring::AbstractString, taxlevel::Union{Int, Symbol})

Finds given taxonomic level in a string (as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria...")) and returns the clade and taxonomic level as a Taxon. If taxon level not given, function will return the most specific (lowest) taxonomic level available.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", 2)
Taxon("Euryarchaeota", :phylum)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", :kingdom)
Taxon("Archaea", :kingdom)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria")
Taxon("Methanobacteria", :class)