MetaPhlAn

BiobakeryUtils.metaphlan_profile — Function

metaphlan_profile(path::AbstractString, level::Union{Int, Symbol}=:all; sample::AbstractString=basename(first(splitext(path))))

Compiles a MetaPhlAn file into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Place name of the CommunityProfile can be specified by passing a sample argument. If name not given, the name of the file becomes the Place name.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profile("test/files/metaphlan_single2.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 96 things in 1 places

Thing names:
Bacteria, Archaea, Firmicutes...Ruminococcus_bromii, Bacteroides_vulgatus

Place names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", 4)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 11 things in 1 places

Thing names:
Clostridiales, Bacteroidales, Coriobacteriales...Firmicutes_unclassified, Pasteurellales

Place names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 things in 1 places

Thing names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Place names:
metaphlan_single2

julia> metaphlan_profile("test/files/metaphlan_single2.tsv", :genus, sample = "sample2")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 40 things in 1 places

Thing names:
Prevotella, Roseburia, Faecalibacterium...Haemophilus, Lactococcus

Place names:
sample2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(paths::Array{<:AbstractString, 1}, level::Union{Int, Symbol}=:all)

Compiles MetaPhlAn profiles from multiple single tables into a CommunityProfile.

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 129 things in 2 places

Thing names:
Bacteria, Firmicutes, Bacteroidetes...Coprococcus_eutactus, Ruminococcus_bromii

Place names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 46 things in 2 places

Thing names:
Bacteroides, Roseburia, Faecalibacterium...Ruthenibacterium, Haemophilus

Place names:
metaphlan_single1, metaphlan_single2

julia> metaphlan_profiles(["test/files/metaphlan_single1.tsv", "test/files/metaphlan_single2.tsv"], 5)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 24 things in 2 places

Thing names:
Lachnospiraceae, Ruminococcaceae, Bacteroidaceae...Clostridiales_unclassified, Pasteurellaceae

Place names:
metaphlan_single1, metaphlan_single2

BiobakeryUtils.metaphlan_profiles — Function

metaphlan_profiles(path::AbstractString, level::Union{Int, Symbol}=:all; keepunidentified=false)

Compiles MetaPhlAn profiles from a merged table into a CommunityProfile. Can select data according to taxonomic level. If level not given, all data is compiled. Set keepunidentified flag to true to keep UNIDENTIFIED data.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples
≡≡≡≡≡≡≡≡≡≡
julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv")
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 42 things in 7 places

Thing names:
Archaea, Euryarchaeota, Methanobacteria...Actinomyces_viscosus, GCF_000175315

Place names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", :genus)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 3 things in 7 places

Thing names:
Methanobrevibacter, Methanosphaera, Actinomyces

Place names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

julia> metaphlan_profiles("test/files/metaphlan_multi_test.tsv", 3)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 2 things in 7 places

Thing names:
Methanobacteria, Actinobacteria

Place names:
sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

# julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv")
# CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 things in 7 places

# Thing names:
# UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

# Place names:
# sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

# julia> metaphlan_profiles("test/files/metaphlan_multi_test_unidentified.tsv", keepunidentified = true)
# CommunityProfile{Float64, Taxon, MicrobiomeSample} with 43 things in 7 places

# Thing names:
# UNIDENTIFIED, Archaea, Euryarchaeota...Actinomyces_viscosus, GCF_000175315

# Place names:
# sample1_taxonomic, sample2_taxonomic, sample3_taxonomic...sample6_taxonomic, sample7_taxonomic

BiobakeryUtils.parsetaxa — Method

parsetaxa(taxstring::AbstractString; throw_error=true)

Given a string representing taxonmic levels as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria..."), separates taxonomic levels into elements of type Taxon in a vector.

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxa("k__Archaea|p__Euryarchaeota|c__Methanobacteria"; throw_error = true)
3-element Vector{Taxon}:
 Taxon("Archaea", :kingdom)
 Taxon("Euryarchaeota", :phylum)
 Taxon("Methanobacteria", :class)

BiobakeryUtils.parsetaxon — Method

parsetaxon(taxstring::AbstractString, taxlevel::Union{Int, Symbol})

Finds given taxonomic level in a string (as formatted by MetaPhlAn (eg "kBacteria|pProteobacteria...")) and returns the clade and taxonomic level as a Taxon. If taxon level not given, function will return the most specific (lowest) taxonomic level available.

Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Examples ≡≡≡≡≡≡≡≡≡≡

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", 2)
Taxon("Euryarchaeota", :phylum)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria", :kingdom)
Taxon("Archaea", :kingdom)

julia> parsetaxon("k__Archaea|p__Euryarchaeota|c__Methanobacteria")
Taxon("Methanobacteria", :class)

BiobakeryUtils.taxfilter! — Function

taxfilter!(df::DataFrame, level::Union{Int, Symbol}; keepunidentified::Bool)

Filter a MetaPhlAn table (as DataFrame) to a particular taxon level. Levels may be given either as numbers or symbols:

1 = :kingdom

2 = :phylum

3 = :class

4 = :order

5 = :family

6 = :genus

7 = :species

8 = :subspecies

Taxon level is removed from resulting taxon string, eg. g__Bifidobacterium becomes Bifidobacterium.

Set keepunidentified flag to false to remove UNIDENTIFIED rows.

taxfilter!() modifies the dataframe that you pass to it and taxfilter() doesn't.

This function will also rename the taxa in the first column.

Examples
≡≡≡≡≡≡≡≡≡≡
julia> df
4×2 DataFrame
 Row │ taxon                          abundance 
     │ String                         Float64   
─────┼──────────────────────────────────────────
   1 │ k__Bacteria                     100.0
   2 │ k__Bacteria|p__Firmicutes        63.1582
   3 │ k__Bacteria|p__Bacteroidetes     25.6038
   4 │ k__Bacteria|p__Actinobacteria    11.0898

julia> taxfilter(df,2; keepunidentified=true)
4×2 DataFrame
 Row │ taxon            abundance 
     │ String           Float64   
─────┼────────────────────────────
   1 │ Firmicutes         63.1582
   2 │ Bacteroidetes      25.6038
   3 │ Actinobacteria     11.0898
   4 │ Verrucomicrobia     0.1482

julia> df
4×2 DataFrame
 Row │ taxon                          abundance 
     │ String                         Float64   
─────┼──────────────────────────────────────────
   1 │ k__Bacteria                     100.0
   2 │ k__Bacteria|p__Firmicutes        63.1582
   3 │ k__Bacteria|p__Bacteroidetes     25.6038
   4 │ k__Bacteria|p__Actinobacteria    11.0898

julia> taxfilter!(df,2; keepunidentified=true)
3×2 DataFrame
 Row │ taxon           abundance 
     │ String          Float64   
─────┼───────────────────────────
   1 │ Firmicutes        63.1582
   2 │ Bacteroidetes     25.6038
   3 │ Actinobacteria    11.0898

julia> df
3×2 DataFrame
 Row │ taxon           abundance 
     │ String          Float64   
─────┼───────────────────────────
   1 │ Firmicutes        63.1582
   2 │ Bacteroidetes     25.6038
   3 │ Actinobacteria    11.0898

julia> taxfilter!(df,1; keepunidentified=true)
3×2 DataFrame
 Row │ taxon           abundance 
     │ String          Float64   
─────┼───────────────────────────
   1 │ Firmicutes        63.1582
   2 │ Bacteroidetes     25.6038
   3 │ Actinobacteria    11.0898