Microbiome.jl functions

BiobakeryUtils.jl re-exports all of the functionality from Microbiome.jl. The docstrings from that package are reproduced here, but checkout the Microbiome.jl docs for more details.

Microbiome.CommunityProfile — Type

CommunityProfile{T, F, S} <: AbstractAbundanceTable{T, F, S}

An AbstractAssemblage from EcoBase.jl that uses an AxisArray of a SparseMatrixCSC under the hood.

CommunityProfiles are tables with AbstractFeature-indexed rows and AbstractSample-indexed columns. Note - we can use the name of samples and features to index.

julia> txs = [Taxon("taxon$i") for i in 1:10];

julia> mss = [MicrobiomeSample("sample$i") for i in 1:5];

julia> mat = spzeros(10,5);

julia> for i in 1:5; mat[i,i] = 1.; end

julia> comm = CommunityProfile(mat, txs, mss)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 things in 5 places

Thing names:
taxon1, taxon2, taxon3...taxon9, taxon10

Place names:
sample1, sample2, sample3, sample4, sample5

julia> comm["taxon1", "sample1"]
1.0

julia> comm[:,["sample1", "sample5"]]
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 things in 2 places

Thing names:
taxon1, taxon2, taxon3...taxon9, taxon10

Place names:
sample1, sample5

julia> comm[Taxon("taxon3", :kingdom), "sample1"]
0.0

Microbiome.GeneFunction — Type

GeneFunction(name::String, taxon::Union{Taxon, String, Missing}) <: AbstractFeature
GeneFunction(name::String)

Microbial gene function object with optional stratification (taxon).

Microbiome.MicrobiomeSample — Type

MicrobiomeSample(name::String, metadata::Dictionary{Symbol, T}) <: AbstractSample
MicrobiomeSample(name::String)

Microbiome sample type that includes a name and a Dictionary of arbitrary metadata using Symbols (other than :name or :metadata) as keys.

Metadata can be accessed using getproperty or getindex on the sample itself.

julia> ms = MicrobiomeSample("sample1", Dictionary([:gender, :age], ["female", 180]))
MicrobiomeSample("sample1", {:gender │ "female", :age │ 180})

julia> name(ms)
"sample1"

julia> ms.name
"sample1"

julia> ms.gender
"female"

julia> ms.age
180

Samples can be instantiated with only a name, leaving the metadata Dictionary blank

julia> ms2 = MicrobiomeSample("sample2")
MicrobiomeSample("sample2", {})

Adding or changing metadata follows the same rules as for the normal Dictionary type.

to change a value use setproperty or setindex. Note that this will fail if the key does not already exist.

to add or remove a value with validation that it does or does not exist already, use insert! and delete! respectively.

to add or remove a value without validation ("upsert"), use set! and unset! respectively.

julia> ms
MicrobiomeSample("sample1", {:gender │ "female", :age │ 180})

julia> ms.age = 16 * 365 # or `ms[:age] = 16 * 365`
5840

julia> set!(ms, :gender, "nonbinary")
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

julia> insert!(ms, :occupation, "clerk")
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840, :occupation │ "clerk"})

julia> insert!(ms, :occupation, "bagger")
ERROR: IndexError("Dictionary already contains index: occupation")

julia> delete!(ms, :occupation)
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

julia> delete!(ms, :occupation)
ERROR: IndexError("Index doesn't exist: occupation")

julia> unset!(ms, :occupation)
MicrobiomeSample("sample1", {:gender │ "nonbinary", :age │ 5840})

Microbiome.Taxon — Type

Taxon(name::String, clade::Union{Missing, Symbol, Int}) <: AbstractFeature
Taxon(name::String)

Microbial taxon with a name and a clade that can be one of

:domain

:kingom

:phylum

:class

:order

:faamily

:genus

:species

:subspecies

:strain

or missing. Contructors can also use numbers 0-9, or pass a string alone (in which case the taxon will be stored as missing)

Microbiome.abundances — Function

abundances(at::AbstractAbundanceTable)

Get the underlying sparse matrix of an AbstractAbundanceTable. Note that this does not copy - any modifications to this matrix will update the parent.

Microbiome.braycurtis — Method

braycurtis(abt::AbstractAbundanceTable)

Returns a pairwise Bray-Curtis dissimilarity matrix.

Microbiome.clade — Method

clade(t::Union{Taxon, missing})

Get the clade field from an Taxon. Returns missing if the clade is not set.

Microbiome.commjoin — Method

commjoin(c1::CommunityProfile, comms::CommunityProfile...)

Join multiple CommunityProfiles, creating a new CommunityProfile. For now, sample names cannot overlap in any of the input profiles.

julia> mss = [MicrobiomeSample(string("sample",i)) for i in 1:15];

julia> txs = [Taxon(string("taxon",i)) for i in 1:20];

julia> cm1 = CommunityProfile(spzeros(10,5), txs[1:10], mss[1:5])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 things in 5 places

Thing names:
taxon1, taxon2, taxon3...taxon9, taxon10

Place names:
sample1, sample2, sample3, sample4, sample5

julia> cm2 = CommunityProfile(spzeros(10,5), txs[6:15], mss[6:10])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 things in 5 places

Thing names:
taxon6, taxon7, taxon8...taxon14, taxon15

Place names:
sample6, sample7, sample8, sample9, sample10

julia> cm3 = CommunityProfile(spzeros(10,5), txs[11:20], mss[11:15])
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 things in 5 places

Thing names:
taxon11, taxon12, taxon13...taxon19, taxon20

Place names:
sample11, sample12, sample13, sample14, sample15

julia> commjoin(cm1, cm2, cm3)
CommunityProfile{Float64, Taxon, MicrobiomeSample} with 20 things in 15 places

Thing names:
taxon1, taxon2, taxon3...taxon19, taxon20

Place names:
sample1, sample2, sample3...sample14, sample15

Microbiome.featurenames — Function

featurenames(at::AbstractAbundanceTable)

Get a vector of feature names from at, equivalent to name.(features(at))

Microbiome.features — Method

features(at::AbstractAbundanceTable)

Returns features in at. To get featurenames instead, use featurenames.

Microbiome.featuretotals — Method

featuretotals(at::AbstractAbundanceTable)

Returns sum of each row (feature) in at. Note, return value is a nfeatures x 1 Matrix, not a Vector. If you need 1D Vector, use vec(featuretotals(at)).

Microbiome.ginisimpson! — Method

ginisimpson!(abt::AbstractAbundanceTable; overwrite=false)

Adds a :ginisimpson entry to the metadata for each sample in abt with the Gini-Simpson alpha diversity of that sample (see ginisimpson). If overwrite=false (the default), uses insert! to perform this operation, so an error will be thrown if any sample already contains a :ginisimpson entry. Otherwise, uses set!.

Microbiome.ginisimpson — Method

ginisimpson(v::Union{AbstractVector, AbstractSparseMatrix}) 
ginisimpson(abt::AbstractAbundanceTable, overwrite=false)

Computes the Gini-Simpson alpha diversity metric for a vector. When called on an AbstractAbundanceTable, returns a 1 x nsamples matrix with 1 entry per sample. See also ginisimpson!.

Microbiome.hasclade — Method

hasclade(t::Taxon)::Bool

Pretty self-explanatory.

Microbiome.hastaxon — Method

hastaxon(t::GeneFunction)::Bool

Pretty self-explanatory.

Microbiome.hellinger — Method

hellinger(abt::AbstractAbundanceTable)

Returns a pairwise Hellinger distance matrix.

Microbiome.jaccard — Method

jaccard(abt::AbstractAbundanceTable)

Returns a pairwise Jaccard distance matrix.

Microbiome.metadata — Method

metadata(t::AbstractSample)

Get the metadata field from an AbstractSample. Note that this is not a copy, so modifications to the returned value will update the parent AbstractSample as well.

Microbiome.metadata — Method

metadata(cp::CommunityProfile)

Returns iterator of NamedTuple per sample, where keys are :sample and each metadata key found in cp. Samples without given metadata are filled with missing.

Returned values can be passed to any Tables.rowtable - compliant type, eg DataFrame.

Microbiome.name — Method

name(t::Union{AbstractSample, AbstractFeature})

Get the name field from an AbstractSample or AbstractFeature.

Microbiome.pcoa — Function

pcoa(abt::AbstractAbundanceTable, f=braycurtis)

Returns eigenvectors from fitting MDS to a distance metric generated by f, by default braycurtis.

Microbiome.present — Function

present(t::Union{Real, Missing}, minabundance::Real=0.0)
present(at::AbstractAbundanceTable, minabundance::Real=0.0)

Check if a given (non-zero) value is greater than or equal to a minimum value. If the minimum abundance is 0, just checks if value is non-zero.

If used on an AbstractAbundanceTable, returns a sparse boolean matrix of the same size.

Microbiome.prevalence — Function

prevalence(a::AbstractArray{<:Real}, minabundance::Real=0.0)
prevalence(at::AbstractAbundanceTable, minabundance::Real=0.0)

Return the fraction of values that are greater than or equal to a minimum. If the minimum abundance is 0, returns the fraction of non-zero values.

If used on an AbstractAbundanceTable, returns a prevalence value for each feature accross the samples.

Microbiome.prevalence_filter — Method

prevalence_filter(comm::AbstractAbundanceTable; minabundance=0.0; minprevalence=0.05, renorm=false)

Return a filtered CommunityProfile where features with prevalence lower than minprevalence are removed. By default, a feature is considered "present" if > 0, but this can be changed by setting minabundance.

Optionally, set renorm = true to calculate relative abundances after low prevalence features are removed.

julia> comm = CommunityProfile(sparse([3 0 1 # 0.33, assuming minabundance 2
                                       2 2 2 # 1.0
                                       0 0 1 # 0.0
                                       2 0 0 # 0.33
                                       ]),
                               [Taxon(string(i)) for i in 1:4],
                               [MicrobiomeSample(string(i)) for i in 1:3]);

julia> prevalence_filter(comm, minabundance=2, minprevalence=0.3) 
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 3 things in 3 places

Thing names:
1, 2, 4

Place names:
1, 2, 3

julia> prevalence_filter(comm, minabundance=2, minprevalence=0.4)
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 1 things in 3 places

Thing names:
2

Place names:
1, 2, 3

julia> prevalence_filter(comm, minabundance=3, minprevalence=0.3)
CommunityProfile{Int64, Taxon, MicrobiomeSample} with 1 things in 3 places

Thing names:
1

Place names:
1, 2, 3

Microbiome.relativeabundance — Function

relativeabundance(at::AbstractAbundanceTable, kind::Symbol=:fraction)

Like relativeabundance!, but does not mutate original.

Microbiome.relativeabundance! — Method

relativeabundance!(a::AbstractAbundanceTable; kind::Symbol=:fraction)

Normalize each sample in AbstractAbundanceTable to the sum of the sample.

By default, columns sum to 1.0. Use kind=:percent for columns to sum to 100.

Microbiome.samplenames — Function

samplenames(at::AbstractAbundanceTable)

Get a vector of sample names from at, equivalent to name.(samples(at))

Microbiome.samples — Method

samples(at::AbstractAbundanceTable)

Returns samples in at. To get samplenames instead, use samplenames.

Microbiome.sampletotals — Method

sampletotals(at::AbstractAbundanceTable)

Returns sum of each row (feature) in at. Note, return value is a 1 x nsamples Matrix, not a Vector. If you need 1D Vector, use vec(sampletotals(at)).

Microbiome.shannon! — Method

shannon!(abt::AbstractAbundanceTable; overwrite=false)

Adds a :shannon entry to the metadata for each sample in abt with the Shannon alpha diversity of that sample (see shannon). If overwrite=false (the default), uses insert! to perform this operation, so an error will be thrown if any sample already contains a :shannon entry. Otherwise, uses set!.

Microbiome.shannon — Method

shannon(v::Union{AbstractVector, AbstractSparseMatrix}) 
shannon(abt::AbstractAbundanceTable)

Computes the Shannon alpha diversity metric for a vector. When called on an AbstractAbundanceTable, returns a 1 x nsamples matrix with 1 entry per sample. See also shannon!.

Microbiome.taxon — Method

taxon(t::Union{GeneFunction, missing})

Get the taxon field from a GeneFunction. Returns missing if the taxon is not set.